Concept Network- Get Unique Nodes

Takes at least two tables of nodes and pagerank (output of `fst_cn_nodes`) and finds nodes unique to one table.

Usage

fst_cn_get_unique(table1, table2, ...)

Arguments

table1

The first table

table2

The second table

...

Any other tables you want to include

Value

Dataframe of words and whether word is unique or not

Examples

con1 <- "elintaso, köyhä, ihminen"
data1 <- conllu_dev_q11_1_nltk
e1 <- fst_cn_edges(data = data1, concepts = con1, threshold = 3)
q11_1_nodes <- fst_cn_nodes(conllu_dev_q11_1_nltk, e1)

con2 <- "kehitysmaa, auttaa, pyrkiä, maa, ihminen"
e2 <- fst_cn_edges(conllu_dev_q11_2_nltk, con2, threshold = 5)
q11_2_nodes <- fst_cn_nodes(conllu_dev_q11_2_nltk, e2)

con3 <- "köyhyys, nälänhätä, sota, ilmastonmuutos, puute"
e3 <- fst_cn_edges(conllu_dev_q11_3_nltk, con3, threshold = 2)
q11_3_nodes <- fst_cn_nodes(conllu_dev_q11_3_nltk, e3)

fst_cn_get_unique(q11_1_nodes, q11_2_nodes)
#>  [1] "alapuolella" "asua"        "auttaa"      "elintaso"    "elää"       
#>  [6] "kehittyä"    "kehitysmaa"  "köyhyysraja" "köyhä"       "köys"       
#> [11] "nähdä"       "nälkä"       "olo"         "osa"         "parantaa"   
#> [16] "pyrkiä"      "saavuttaa"   "suuri"       "tarvita"     "tarvitsla"  
#> [21] "tilanne"     "voida"      
fst_cn_get_unique(q11_1_nodes, q11_2_nodes, q11_3_nodes)
#>  [1] "aiheuttaa"         "alapuolella"       "asua"             
#>  [4] "auttaa"            "elintaso"          "elää"             
#>  [7] "epätasa-arvo"      "huono"             "ihmisoikeus"      
#> [10] "ilmastonmuutos"    "kehittyä"          "kehitysmaa"       
#> [13] "koulutus"          "köyhyys"           "köyhyysraja"      
#> [16] "köyhä"             "köys"              "nähdä"            
#> [19] "nälänhätä"         "olo"               "osa"              
#> [22] "parantaa"          "puhdas"            "puute"            
#> [25] "pyrkiä"            "riippuvuus"        "saastuminen"      
#> [28] "saavuttaa"         "sota"              "suuri"            
#> [31] "talous"            "tarvita"           "tarvitsla"        
#> [34] "tilanne"           "vesi"              "voida"            
#> [37] "väestönkasvu"      "ylikansoittuminen" "öljy"