Concept Network- Get unique nodes from separate top n-grams tables

Takes at least two tables of nodes and pagerank (output of `fst_cn_nodes()`) and finds nodes unique to one table.

Usage

fst_cn_get_unique_separate(table1, table2, ...)

Arguments

table1

The first table.

table2

The second table.

...

Any other tables you want to include.

Value

Dataframe of words and whether word is unique or not.

Examples

pos_filter <- c("NOUN", "VERB", "ADJ", "ADV")
e1 <- fst_cn_edges(fst_child, "lyödä", pos_filter = pos_filter)
e2 <- fst_cn_edges(fst_child, "lyöminen", pos_filter = pos_filter)
n1 <- fst_cn_nodes(fst_child, e1)
n2 <- fst_cn_nodes(fst_child, e2)
fst_cn_get_unique_separate(n1, n2)
#> [1] "lyödä"      "lyöminen"   "potkia"     "potkiminen" "sanoa"     
#> [6] "toinen"     "töniminen"