Get unique n-grams

Takes at least two tables of n-grams and frequencies (either output of `fst_get_top_words()` or `fst_get_top_ngrams()`) and finds n-grams unique to one table.

Usage

fst_get_unique_ngrams(table1, table2, ...)

Arguments

table1

The first table.

table2

The second table.

...

Any other tables you want to include.

Value

Dataframe of words and whether word is unique or not.

Examples

top_f <- fst_get_top_words(conllu_dev_q11_1_f_nltk)
#> Note:
#>  Words with equal occurrence are presented in alphabetial order. 
#>  By default, words are presented in order to the `number` cutoff word. 
#>  This means that equally-occurring later-alphabetically words beyond the cutoff word will not be displayed.
#> 
top_m <- fst_get_top_words(conllu_dev_q11_1_m_nltk)
#> Note:
#>  Words with equal occurrence are presented in alphabetial order. 
#>  By default, words are presented in order to the `number` cutoff word. 
#>  This means that equally-occurring later-alphabetically words beyond the cutoff word will not be displayed.
#> 
top_na <- fst_get_top_words(conllu_dev_q11_1_na_nltk)
#> Note:
#>  Words with equal occurrence are presented in alphabetial order. 
#>  By default, words are presented in order to the `number` cutoff word. 
#>  This means that equally-occurring later-alphabetically words beyond the cutoff word will not be displayed.
#> 
topn_f <- fst_get_top_ngrams(conllu_dev_q11_1_f_nltk)
#> Note:
#>  N-grams with equal occurrence are presented in alphabetial order. 
#>  By default, n-grams are presented in order to the `number` cutoff n-gram. 
#>  This means that equally-occurring later-alphabetically n-grams beyond the cutoff n-gram will not be displayed. 
#> 
topn_m <- fst_get_top_ngrams(conllu_dev_q11_1_m_nltk)
#> Note:
#>  N-grams with equal occurrence are presented in alphabetial order. 
#>  By default, n-grams are presented in order to the `number` cutoff n-gram. 
#>  This means that equally-occurring later-alphabetically n-grams beyond the cutoff n-gram will not be displayed. 
#> 
topn_na <- fst_get_top_ngrams(conllu_dev_q11_1_na_nltk)
#> Note:
#>  N-grams with equal occurrence are presented in alphabetial order. 
#>  By default, n-grams are presented in order to the `number` cutoff n-gram. 
#>  This means that equally-occurring later-alphabetically n-grams beyond the cutoff n-gram will not be displayed. 
#> 
fst_get_unique_ngrams(top_f, top_m, top_na)
#> # A tibble: 15 × 2
#>    words    unique_word
#>    <fct>    <chr>      
#>  1 maa      no         
#>  2 apu      no         
#>  3 köyhyys  no         
#>  4 tarvita  yes        
#>  5 elintaso no         
#>  6 köyhä    no         
#>  7 huono    no         
#>  8 paljon   no         
#>  9 asia     no         
#> 10 ihminen  no         
#> 11 matala   yes        
#> 12 alhainen yes        
#> 13 kehitys  no         
#> 14 elinolo  yes        
#> 15 ruoka    yes        
fst_get_unique_ngrams(topn_f, topn_m, topn_na)
#> # A tibble: 15 × 2
#>    words    unique_word
#>    <fct>    <chr>      
#>  1 maa      no         
#>  2 apu      no         
#>  3 köyhyys  no         
#>  4 tarvita  yes        
#>  5 elintaso no         
#>  6 köyhä    no         
#>  7 huono    no         
#>  8 paljon   no         
#>  9 asia     no         
#> 10 ihminen  no         
#> 11 matala   yes        
#> 12 alhainen yes        
#> 13 kehitys  no         
#> 14 elinolo  yes        
#> 15 ruoka    yes