Compare response lengths

Compare length of text responses for between 2 and 4 sets of prepared data.

Usage

fst_length_compare(
  data1,
  data2,
  data3 = NULL,
  data4 = NULL,
  name1 = "Group 1",
  name2 = "Group 2",
  name3 = "Group 3",
  name4 = "Group 4",
  incl_sentences = TRUE
)

Arguments

data1

A dataframe of text in CoNLL-U format for the first group.

data2

A dataframe of text in CoNLL-U format for the second group.

data3

An optional dataframe of text in CoNLL-U format for the third group, default is `NULL`.

data4

An optional dataframe of text in CoNLL-U format for the fourth group, default is `NULL`.

name1

A string describing data1, default is `"Group 1"`.

name2

A string describing data2, default is `"Group 2"`.

name3

A string describing data3, default is `"Group 3"`.

name4

A string describing data4, default is `"Group 4"`.

incl_sentences

Whether to include sentence data in table, default is `TRUE`.

Value

Dataframe summarising response lengths.

Examples

f <- conllu_dev_q11_1_f_nltk
m <- conllu_dev_q11_1_m_nltk
na <- conllu_dev_q11_1_na_nltk
all <- conllu_dev_q11_1_nltk
fst_length_compare(f, m, na, all, "Female", "Male", "Not Spec", "All")
#> # A tibble: 8 × 8
#>   Description         Respondents  Mean Minimum    Q1 Median    Q3 Maximum
#>   <chr>                     <int> <dbl>   <int> <dbl>  <dbl> <dbl>   <int>
#> 1 Female- Words               662  7.00       1     4    6       9      33
#> 2 Female- Sentences           662  1.06       1     1    1       1       4
#> 3 Male- Words                 175  5.15       1     3    4       6      27
#> 4 Male- Sentences             175  1.03       1     1    1       1       2
#> 5 Not Spec- Words              84  5.57       1     3    4.5     7      19
#> 6 Not Spec- Sentences          84  1.06       1     1    1       1       3
#> 7 All- Words                  921  6.51       1     3    5       8      33
#> 8 All- Sentences              921  1.06       1     1    1       1       4
fst_length_compare(f, m, name1 = "F", name2 = "M", incl_sentences = FALSE)
#> # A tibble: 2 × 8
#>   Description Respondents  Mean Minimum    Q1 Median    Q3 Maximum
#>   <chr>             <int> <dbl>   <int> <dbl>  <dbl> <dbl>   <int>
#> 1 F- Words            662  7.00       1     4      6     9      33
#> 2 M- Words            175  5.15       1     3      4     6      27