Make comparison summary

Compare text responses for different groups of participants. Data is split based on different values in the `field` column of formatted data. Results will be shown within the plots pane.

Usage

fst_summarise_compare(
  data,
  field,
  exclude_nulls = FALSE,
  rename_nulls = "null_data"
)

Arguments

data

A dataframe of text in CoNLL-U format with additional `field` column for splitting data.

field

Column in `data` used for splitting groups

exclude_nulls

Whether to include NULLs in `field` column, default is `FALSE`

rename_nulls

What to fill NULL values with if `exclude_nulls = FALSE`.

Value

Summary table of responses between groups.

Examples

fst_summarise_compare(fst_child, 'gender')
#>    Description Respondents No Response Proportion Total Words Unique Words
#>         <char>       <int>       <int>      <num>       <int>        <int>
#> 1:         boy         212           0          1         721          334
#> 2:        girl         201           0          1         859          357
#>    Unique Lemmas
#>            <int>
#> 1:           252
#> 2:           270
fst_summarise_compare(fst_dev_coop, 'gender')
#>    Description Respondents No Response Proportion Total Words Unique Words
#>         <char>       <int>       <int>      <num>       <int>        <int>
#> 1:      Female         673          13       0.98        2993          823
#> 2:        Male         183           8       0.96         795          383
#> 3:   null_data          89           4       0.96         404          225
#>    Unique Lemmas
#>            <int>
#> 1:           722
#> 2:           354
#> 3:           208